Refactor service to filter
