--- src/Language/en/Visits.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Tatter\Visits\Language\en;
-
-return [
-    'noTrackingMethod'     => 'No tracking method selected.',
-    'invalidResetMinutes'  => 'Minutes-to-reset must be a positive integer or zero.',
-    'missingDatabaseTable' => 'Table "{0}" missing for visit storage.',
-];
